refactor(LottoPurchase):
- LottoHelper -> LottoPurchase 클래스명 변경
- setter 삭제 (생성자로 이동)
- 관련 클래스에 반영
