+++ src/main/java/study/lotto/LottoGame.java
@@ -1,6 +1,6 @@
 package study.lotto;
 
-import study.lotto.helper.LottoHelper;
+import study.lotto.model.LottoPurchase;
 import study.lotto.model.Lottos;
 import study.lotto.service.LottoService;
 import study.lotto.view.InputView;
@@ -8,13 +8,11 @@ import study.lotto.view.InputView;
 public class LottoGame {
 
     public static void main(String[] args) {
-        LottoHelper lottoHelper = new LottoHelper(InputView.scanPurchaseAmount());
-        lottoHelper.setManualLottoCount(InputView.scanManualLottoCount());
-
+        LottoPurchase lottoPurchase = new LottoPurchase(InputView.scanPurchaseAmount(), InputView.scanManualLottoCount());
         LottoService lottoService = new LottoService();
 
-        Lottos lottos = lottoService.purchaseLottos(lottoHelper);
+        Lottos lottos = lottoService.purchaseLottos(lottoPurchase);
 
-        lottoService.checkLottoResult(lottos, lottoHelper);
+        lottoService.checkLottoResult(lottos, lottoPurchase);
     }
 }
