+++ src/main/java/study/lotto/service/LottoService.java
@@ -1,8 +1,8 @@
 package study.lotto.service;
 
-import study.lotto.helper.LottoHelper;
 import study.lotto.model.LottoResult;
 import study.lotto.model.Lottos;
+import study.lotto.model.LottoPurchase;
 import study.lotto.model.WinningLottoInfo;
 import study.lotto.utils.StatisticUtils;
 import study.lotto.view.InputView;
@@ -10,14 +10,14 @@ import study.lotto.view.ResultView;
 
 public class LottoService {
 
-    public Lottos purchaseLottos(LottoHelper lottoHelper) {
-        Lottos lottos = purchaseManualLottos(lottoHelper.getManualLottoCount());
+    public Lottos purchaseLottos(LottoPurchase lottoPurchase) {
+        Lottos lottos = purchaseManualLottos(lottoPurchase.getManualLottoCount());
 
-        lottos.addAll(purchaseAutoLottos(lottoHelper.getAutoLottoCount()));
+        lottos.addAll(purchaseAutoLottos(lottoPurchase.getAutoLottoCount()));
 
         ResultView.printPurchaseMessage(
-                lottoHelper.getManualLottoCount(),
-                lottoHelper.getAutoLottoCount(),
+                lottoPurchase.getManualLottoCount(),
+                lottoPurchase.getAutoLottoCount(),
                 lottos
         );
 
@@ -32,11 +32,11 @@ public class LottoService {
         return Lottos.of(autoLottoCount);
     }
 
-    public void checkLottoResult(Lottos lottos, LottoHelper lottoHelper) {
+    public void checkLottoResult(Lottos lottos, LottoPurchase lottoPurchase) {
         WinningLottoInfo winningLottoInfo = WinningLottoInfo.of(InputView.scanWinningNumbers(), InputView.scanBonusNumber());
 
         LottoResult lottoResult = LottoResult.produce(lottos, winningLottoInfo);
 
-        ResultView.printStatisticsMessage(lottoResult, StatisticUtils.calculateEarningRate(lottoHelper.getPurchaseAmount(), lottoResult));
+        ResultView.printStatisticsMessage(lottoResult, StatisticUtils.calculateEarningRate(lottoPurchase.getPurchaseAmount(), lottoResult));
     }
 }
